+++ asdf_astropy/converters/coordinates/angle.py
@@ -6,6 +6,7 @@ class AngleConverter(QuantityConverter):
     types = (
         "astropy.coordinates.angles.Angle",
         "astropy.coordinates.angles.core.Angle",
+        "astropy.utils.masked.core.MaskedAngle",
     )
 
     def from_yaml_tree(self, node, tag, ctx):
@@ -19,6 +20,7 @@ class LatitudeConverter(QuantityConverter):
     types = (
         "astropy.coordinates.angles.Latitude",
         "astropy.coordinates.angles.core.Latitude",
+        "astropy.utils.masked.core.MaskedLatitude",
     )
 
     def from_yaml_tree(self, node, tag, ctx):
@@ -32,6 +34,7 @@ class LongitudeConverter(QuantityConverter):
     types = (
         "astropy.coordinates.angles.Longitude",
         "astropy.coordinates.angles.core.Longitude",
+        "astropy.utils.masked.core.MaskedLongitude",
     )
 
     def to_yaml_tree(self, obj, tag, ctx):